requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
protobuf==3.11.2
psutil==5.6.7
python-bitcoinrpc==1.0
requests==2.22.0
rope==0.14.0
six==1.13.0
urllib3==1.25.7
